Add rng controller and finish round actions

+++ Rlatro.Core/GameEngine/GameStateController/PersistentStates/PersistentState.cs
@@ -6,9 +6,9 @@ namespace Balatro.Core.GameEngine.GameStateController.PersistentStates
     public class PersistentState
     {
         public uint Gold { get; set; }
-        public byte Discards { get; set; }
-        public byte Hands { get; set; }
-        public byte HandSize { get; set; }
+        public int Discards { get; set; }
+        public int Hands { get; set; }
+        public int HandSize { get; set; }
 
         // TODO: Implement the logic for these properties
         public ScoreContext GetHandScore(HandRank rank) => new ScoreContext()
